pyproject.toml
[project]
name = "levibes"
version = "0.1.0"
description = "Motivational Image Caption Generator - Create inspiring social media content with AI-generated captions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "art",
    "gratient",
    "openai>=1.93.0",
    "pillow>=11.3.0",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "yaspin>=3.1.0",
]

[project.scripts]
levibes = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
